style: Structure and rename save-map files
