+++ projects/hslayers/src/components/add-data/catalogue/catalogue.component.ts
@@ -9,7 +9,7 @@ import {HsConfig} from '../../../config.service';
 import {HsCoreService} from '../../core/core.service';
 import {HsEndpoint} from '../../../common/endpoints/endpoint.interface';
 import {HsLanguageService} from '../../language/language.service';
-import {HsLaymanService} from '../../save-map/layman.service';
+import {HsLaymanService} from '../../save-map/feature-services/layman.service';
 import {HsLayoutService} from '../../layout/layout.service';
 import {HsUtilsService} from '../../utils/utils.service';
