+++ projects/hslayers/src/components/add-data/common/add-layer-authorized/add-layer-authorized.component.ts
@@ -4,7 +4,7 @@ import {
   HsAddDataCommonFileService,
   HsAddDataCommonFileServiceParams,
 } from '../common-file.service';
-import {HsLaymanService} from '../../../save-map/layman.service';
+import {HsLaymanService} from '../../../save-map/feature-services/layman.service';
 import {HsUtilsService} from '../../../utils/utils.service';
 import {fileDataObject} from '../../file/types/file-data-object.type';
