+++ projects/hslayers/src/components/add-data/common/common-file.service.ts
@@ -11,8 +11,8 @@ import {HsAddDataService} from '../add-data.service';
 import {HsCommonEndpointsService} from '../../../common/endpoints/endpoints.service';
 import {HsEndpoint} from '../../../common/endpoints/endpoint.interface';
 import {HsLanguageService} from '../../language/language.service';
-import {HsLaymanLayerDescriptor} from '../../save-map/layman-layer-descriptor.interface';
-import {HsLaymanService} from '../../save-map/layman.service';
+import {HsLaymanLayerDescriptor} from '../../save-map/interfaces/layman-layer-descriptor.interface';
+import {HsLaymanService} from '../../save-map/feature-services/layman.service';
 import {HsLogService} from '../../../common/log/log.service';
 import {HsToastService} from '../../layout/toast/toast.service';
 import {PREFER_RESUMABLE_SIZE_LIMIT} from '../../save-map/layman-utils';
