+++ projects/hslayers/src/components/compositions/compositions-catalogue.service.ts
@@ -9,7 +9,7 @@ import {HsCompositionsService} from './compositions.service';
 import {HsDialogContainerService} from '../layout/dialogs/dialog-container.service';
 import {HsEndpoint} from '../../common/endpoints/endpoint.interface';
 import {HsEventBusService} from '../core/event-bus.service';
-import {HsLaymanService} from '../save-map/layman.service';
+import {HsLaymanService} from '../save-map/feature-services/layman.service';
 import {HsLayoutService} from '../layout/layout.service';
 import {HsMapService} from '../map/map.service';
 import {HsUtilsService} from '../utils/utils.service';
