+++ projects/hslayers/src/components/compositions/compositions.service.ts
@@ -20,7 +20,7 @@ import {HsEventBusService} from '../core/event-bus.service';
 import {HsLanguageService} from '../language/language.service';
 import {HsLogService} from '../../common/log/log.service';
 import {HsShareUrlService} from '../permalink/share-url.service';
-import {HsStatusManagerService} from '../save-map/status-manager.service';
+import {HsStatusManagerService} from '../save-map/feature-services/status-manager.service';
 import {HsToastService} from '../layout/toast/toast.service';
 import {HsUtilsService} from '../utils/utils.service';
