+++ projects/hslayers/src/components/draw/draw.service.ts
@@ -22,7 +22,7 @@ import {HsEventBusService} from '../core/event-bus.service';
 import {HsLanguageService} from './../language/language.service';
 import {HsLayerUtilsService} from '../utils/layer-utils.service';
 import {HsLaymanBrowserService} from '../add-data/catalogue/layman/layman.service';
-import {HsLaymanService} from '../save-map/layman.service';
+import {HsLaymanService} from '../save-map/feature-services/layman.service';
 import {HsLayoutService} from '../layout/layout.service';
 import {HsLogService} from '../../common/log/log.service';
 import {HsMapService} from '../map/map.service';
