+++ projects/hslayers/src/components/save-map/dialog-result/dialog-result.component.html
