+++ projects/hslayers/src/components/save-map/dialog-result/dialog-result.component.ts
@@ -1,12 +1,14 @@
-import {Component, ViewRef} from '@angular/core';
-import {HsDialogComponent} from '../layout/dialogs/dialog-component.interface';
-import {HsDialogContainerService} from '../layout/dialogs/dialog-container.service';
-import {HsSaveMapManagerService} from './save-map-manager.service';
+import {Component, OnInit, ViewRef} from '@angular/core';
+
+import {HsDialogComponent} from '../../layout/dialogs/dialog-component.interface';
+import {HsDialogContainerService} from '../../layout/dialogs/dialog-container.service';
+import {HsSaveMapManagerService} from '../feature-services/save-map-manager.service';
 @Component({
   selector: 'hs-save-map-dialog-result',
-  templateUrl: './partials/dialog_result.html',
+  templateUrl: './dialog-result.component.html',
 })
-export class HsSaveMapResultDialogComponent implements HsDialogComponent {
+export class HsSaveMapResultDialogComponent
+  implements HsDialogComponent, OnInit {
   viewRef: ViewRef;
   data: any;
   appRef;
