+++ projects/hslayers/src/components/save-map/interfaces/layman-layer-descriptor.interface.ts
@@ -1,4 +1,4 @@
-import {StatusStateType} from './status-state.type';
+import {StatusStateType} from '../types/status-state.type';
 
 export interface HsLaymanLayerDescriptor {
   uuid?: string;
