+++ projects/hslayers/src/components/save-map/interfaces/saver-service.interface.ts
