+++ projects/hslayers/src/components/save-map/layman-utils.ts
@@ -2,7 +2,7 @@ import {Layer} from 'ol/layer';
 import {Source} from 'ol/source';
 
 import {HsEndpoint} from '../../common/endpoints/endpoint.interface';
-import {HsLaymanLayerDescriptor} from './layman-layer-descriptor.interface';
+import {HsLaymanLayerDescriptor} from './interfaces/layman-layer-descriptor.interface';
 import {getName, getTitle} from '../../common/layer-extensions';
 
 export const PREFER_RESUMABLE_SIZE_LIMIT = 2 * 1024 * 1024; // 2 MB
