+++ projects/hslayers/src/components/save-map/public-api.ts
@@ -1,16 +1,16 @@
-export * from './dialog-spawner.service';
-export * from './layer-synchronizer.service';
-export * from './layman-layer-descriptor.interface';
+export * from './feature-services/dialog-spawner.service';
+export * from './feature-services/layer-synchronizer.service';
+export * from './interfaces/layman-layer-descriptor.interface';
 export * from './layman-utils';
-export * from './layman.service';
-export * from './save-map-form.component';
-export * from './save-map-dialog.component';
-export * from './save-map-manager.service';
+export * from './feature-services/layman.service';
+export * from './form/form.component';
+export * from './dialog-save/dialog-save.component';
+export * from './feature-services/save-map-manager.service';
 export * from './save-map.component';
 export * from './save-map.module';
-export * from './save-map.result-dialog.component';
+export * from './dialog-result/dialog-result.component';
 export * from './save-map.service';
-export * from './saver-service.interface';
-export * from './status-manager.service';
-export * from './status-state.type';
-export * from './bounding-box-object.type';
+export * from './interfaces/saver-service.interface';
+export * from './feature-services/status-manager.service';
+export * from './types/status-state.type';
+export * from './types/bounding-box-object.type';
