+++ projects/hslayers/src/components/save-map/save-map.module.ts
@@ -5,10 +5,10 @@ import {HsLanguageModule} from '../language/language.module';
 import {HsLaymanModule} from '../../common/layman/layman.module';
 import {HsLogModule} from '../../common/log/log.module';
 import {HsPanelHelpersModule} from '../layout/panels/panel-helpers.module';
-import {HsSaveMapAdvancedFormComponent} from './save-map-form.component';
+import {HsSaveMapAdvancedFormComponent} from './form/form.component';
 import {HsSaveMapComponent} from './save-map.component';
-import {HsSaveMapDialogComponent} from './save-map-dialog.component';
-import {HsSaveMapResultDialogComponent} from './save-map.result-dialog.component';
+import {HsSaveMapDialogComponent} from './dialog-save/dialog-save.component';
+import {HsSaveMapResultDialogComponent} from './dialog-result/dialog-result.component';
 import {HsUiExtensionsModule} from '../../common/widgets/ui-extensions.module';
 
 @NgModule({
