+++ projects/hslayers/src/components/save-map/save-map.spec.ts
@@ -19,14 +19,14 @@ import {HsEndpoint} from '../../common/endpoints/endpoint.interface';
 import {HsEventBusService} from '../core/event-bus.service';
 import {HsLanguageModule} from '../language/language.module';
 import {HsLayerUtilsService} from '../utils/layer-utils.service';
-import {HsLaymanService} from './layman.service';
+import {HsLaymanService} from './feature-services/layman.service';
 import {HsLayoutService} from '../layout/layout.service';
 import {HsLayoutServiceMock} from '../layout/layout.service.mock';
 import {HsMapService} from '../map/map.service';
 import {HsMapServiceMock} from '../map/map.service.mock';
 import {HsSaveMapComponent} from './save-map.component';
-import {HsSaveMapDialogSpawnerService} from './dialog-spawner.service';
-import {HsSaveMapManagerService} from './save-map-manager.service';
+import {HsSaveMapDialogSpawnerService} from './feature-services/dialog-spawner.service';
+import {HsSaveMapManagerService} from './feature-services/save-map-manager.service';
 import {HsSaveMapManagerServiceMock} from './save-map-manager.service.mock';
 import {HsUtilsService} from '../utils/utils.service';
 import {HsUtilsServiceMock} from '../utils/utils.service.mock';
