+++ projects/hslayers/src/components/save-map/types/bounding-box-object.type.ts
