+++ projects/hslayers/src/components/save-map/types/status-state.type.ts
